+++ Magna-World/src/main/java/de/morigm/magnaworld/loader/PluginLoader.java
@@ -3,6 +3,7 @@ package de.morigm.magnaworld.loader;
 import de.morigm.magnaworld.Main;
 import de.morigm.magnaworld.commands.MWCREATE;
 import de.morigm.magnaworld.commands.MWDELETE;
+import de.morigm.magnaworld.commands.MWLIST;
 import de.morigm.magnaworld.commands.MWTP;
 
 public class PluginLoader 
@@ -13,7 +14,8 @@ public class PluginLoader
 		new MWCREATE().register("mwcreate", Main.getInstance().getLanguage(), Main.getInstance().getPermission(),Main.getInstance());
 		new MWTP().register("mwtp", Main.getInstance().getLanguage(), Main.getInstance().getPermission(),Main.getInstance());
 		new MWDELETE().register("mwdelete", Main.getInstance().getLanguage(), Main.getInstance().getPermission(),Main.getInstance());
-}
+		new MWLIST().register("mwlist", Main.getInstance().getLanguage(), Main.getInstance().getPermission(),Main.getInstance());
+	}
 	
 	public void loadListener() 
 	{